requirements.txt
biopython==1.66
nltk==3.2.1
